And it builds on CentOS again

+++ SOURCES/ocaml-mirage-xen.install.sh
@@ -13,7 +13,9 @@ mkdir -p $OCAMLFIND_DESTDIR/mirage-xen
 # Why does ocamlfind ignore me and delete that directory??
 MIRAGE_OS=xen ./cmd install
 mkdir -p $OCAMLFIND_DESTDIR/mirage-xen
-cp $2/mirage-xen/* $OCAMLFIND_DESTDIR/mirage-xen
+if [ -e $2/mirage-xen ]; then
+  cp $2/mirage-xen/* $OCAMLFIND_DESTDIR/mirage-xen
+fi
 find . -name "*.a" -exec cp {} $1/mirage-xen \;
 cp _build/runtime/kernel/libxen.a $1/mirage-xen
 cp _build/runtime/kernel/longjmp.o $1/mirage-xen
